third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Ensure correct behavior of drawImage with ImageBitmaps from video elements.");
window.jsTestIsAsync = true;

function jsWrapperClass(node)
{
    // returns the ClassName of node
    if (!node)
        return "[null]";
    var string = Object.prototype.toString.apply(node);

    // string will be of the form [object ClassName]
    return string.substr(8, string.length - 9);
}

function shouldBeType(expression, className)
{
    shouldBe("jsWrapperClass(" + expression + ")", "'" + className + "'");
}

function shouldBeOpaque(x, y) {
    d = ctx.getImageData(x, y, 1, 1).data;
    shouldBeTrue("d[3] == 255");
}

function shouldBeClear(x, y) {
    // should be transparent black pixels
    d = ctx.getImageData(x, y, 1, 1).data;
    shouldBeTrue("d[0] == 0");
    shouldBeTrue("d[1] == 0");
    shouldBeTrue("d[2] == 0");
    shouldBeTrue("d[3] == 0");
}

function clearContext() {
    ctx.clearRect(0, 0, 500, 500);
}

var bitmap;
var video;
// check video cropping with all of these functions
var checks = [checkNoCrop, checkCrop, checkCropRight, checkOverCrop, checkOverCropRight, checkNegativeCrop, checkEmpty, checkEmpty2];

var canvas = document.createElement("canvas");
canvas.setAttribute("width", "500");
canvas.setAttribute("height", "500");
var ctx = canvas.getContext("2d");

video = document.createElement("video");
video.addEventListener("canplaythrough", videoLoaded, false);
video.src = "../../compositing/resources/video.ogv";

function videoLoaded() {
    // wait for callback to finish before each check to ensure synchronous behavior
    nextCheck();
}

var i = 0;
function nextCheck() {
    checks[i](video);
    i += 1;
}

function checkNoCrop(element) {
    debug("checkNoCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackNoCrop);
}

function checkCrop(element) {
    debug("checkCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackCrop, 0, 0, 100, 100);
}

function checkCropCenter(element) {
    debug("checkCropCenter with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackCrop, 50, 50, 150, 150);
}

function checkCropRight(element) {
    debug("checkCropRight with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackCrop, 100, 100, 100, 100);
}

function checkOverCrop(element) {
    debug("checkOverCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackOverCrop, -100, -100, 600, 600);
}

function checkOverCropRight(element) {
    debug("checkOverCropRight with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackOverCropRight, 100, 100, 500, 500);
}

function checkNegativeCrop(element) {
    debug("checkNegativeCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackCrop, 100, 100, -100, -100);
}

function checkEmpty(element) {
    debug("checkNoCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackEmpty, -300, -300, 300, 300);
}

function checkEmpty2(element) {
    debug("checkNoCrop with ".concat(jsWrapperClass(element)));
    createImageBitmap(element, callbackEmpty, 400, 300, 300, 300);
}

function callbackNoCrop(imageBitmap) {
    // should be drawn to (0, 0), (352, 288)
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeOpaque(10, 10);
    shouldBeOpaque(100, 100);
    shouldBeOpaque(300, 50);
    shouldBeOpaque(350, 280);
    shouldBeClear(360, 40);
    shouldBeClear(80, 290);

    // shrunk to (0, 0), (100, 100)
    clearContext(ctx);
    ctx.drawImage(imageBitmap, 0, 0, 100, 100);
    shouldBeOpaque(40, 40);
    shouldBeOpaque(90, 90);
    shouldBeClear(10, 110);
    shouldBeClear(110, 10);
    shouldBeClear(110, 110);

    // shrunk to (100, 100), (250, 200)
    clearContext(ctx);
    ctx.drawImage(imageBitmap, 100, 100, 150, 100);
    shouldBeClear(90, 90);
    shouldBeOpaque(110, 110);
    shouldBeOpaque(240, 190);
    shouldBeClear(110, 210);
    shouldBeClear(260, 110);
    shouldBeClear(260, 210);

    // black should be drawn to (100, 100), (200, 200)
    clearContext(ctx);
    ctx.drawImage(imageBitmap, 100, 100, 100, 100, 100, 100, 100, 100);
    shouldBeClear(90, 90);
    shouldBeOpaque(110, 110);
    shouldBeOpaque(190, 190);
    shouldBeClear(10, 210);
    shouldBeClear(210, 10);
    shouldBeClear(210, 210);

    // nothing should be drawn
    clearContext(ctx);
    ctx.drawImage(imageBitmap, 400, 300, 200, 200, 100, 100, 100, 100);
    shouldBeClear(10, 10);
    shouldBeClear(80, 80);
    shouldBeClear(150, 150);
    shouldBeClear(210, 210);

    commonCallback(imageBitmap);
}

function callbackCrop(imageBitmap) {
    // should be drawn to (0, 0), (100, 100)
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeOpaque(10, 10);
    shouldBeOpaque(90, 90);
    shouldBeClear(110, 110);
    shouldBeClear(210, 210);

    commonCallback(imageBitmap);
}

function callbackOverCrop(imageBitmap) {
    // should be drawn to (100, 100), (452, 388)
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeClear(10, 10);
    shouldBeClear(90, 90);
    shouldBeOpaque(110, 110);
    shouldBeOpaque(300, 200);
    shouldBeOpaque(450, 380);
    shouldBeClear(500, 50);
    shouldBeClear(50, 350);
    shouldBeClear(460, 390);

    // should be drawn to (50, 50), (226, 194)
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0, 300, 300);
    shouldBeClear(10, 10);
    shouldBeClear(40, 40);
    shouldBeOpaque(60, 60);
    shouldBeOpaque(220, 190);
    shouldBeClear(230, 50);
    shouldBeClear(50, 200);
    shouldBeClear(230, 200);

    commonCallback(imageBitmap);
}

function callbackOverCropRight(imageBitmap) {
    // should be drawn to (0, 0), (252, 188)
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeOpaque(10, 10);
    shouldBeOpaque(100, 100);
    shouldBeOpaque(200, 50);
    shouldBeOpaque(50, 150);
    shouldBeOpaque(250, 180);
    shouldBeClear(360, 40);
    shouldBeClear(80, 190);
    shouldBeClear(260, 190);

    commonCallback(imageBitmap);
}

function callbackEmpty(imageBitmap) {
    // nothing should be drawn
    clearContext();
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeClear(10, 10);
    shouldBeClear(90, 90);
    shouldBeClear(110, 110);
    shouldBeClear(210, 210);

    commonCallback(imageBitmap);
}

function commonCallback(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType("bitmap", "ImageBitmap");

    if (i < checks.length) {
        nextCheck();
    } else {
        finishJSTest();
    }
}

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
